better bsp structure basic command scheduling
