+++ bsp/boards/nrf51822_crazyflie2/crazyflie/cf_pm.h
@@ -65,6 +65,6 @@ void pmSysBootloader(bool enable);
 // Lan HUANG (Yelloooblue@outlook.com) June 2024.
 // @brief automatic boot STM32 on Crazyflie.
 
-void crazyflie_init();
+void pm_boot_all();
 
 #endif //__CF_PM_H__
