+++ bsp/boards/nrf51822_crazyflie2/crazyflie/cf_uart.h
@@ -35,7 +35,7 @@
  */
 
 
-#ifndef __CF_ART_H__
+#ifndef __CF_UART_H__
 #define __CF_UART_H__
 
 #include <stdbool.h>
